Flyttet validering til BrevService, lagd funskjon lagHenleggelsesBrevRequest
