+++ src/main/kotlin/no/nav/familie/klage/henlegg/HenleggBehandlingController.kt
@@ -2,12 +2,8 @@ package no.nav.familie.klage.henlegg
 
 import no.nav.familie.klage.brev.BrevService
 import no.nav.familie.klage.felles.domain.AuditLoggerEvent
-import no.nav.familie.klage.felles.util.isEqualOrAfter
-import no.nav.familie.klage.infrastruktur.exception.feilHvis
 import no.nav.familie.klage.infrastruktur.sikkerhet.TilgangService
-import no.nav.familie.klage.personopplysninger.PersonopplysningerService
 import no.nav.familie.kontrakter.felles.Ressurs
-import no.nav.familie.kontrakter.felles.klage.HenlagtÅrsak
 import no.nav.security.token.support.core.api.ProtectedWithClaims
 import org.springframework.validation.annotation.Validated
 import org.springframework.web.bind.annotation.GetMapping
@@ -16,7 +12,6 @@ import org.springframework.web.bind.annotation.PostMapping
 import org.springframework.web.bind.annotation.RequestBody
 import org.springframework.web.bind.annotation.RequestMapping
 import org.springframework.web.bind.annotation.RestController
-import java.time.LocalDate
 import java.util.UUID
 
 @RestController
@@ -27,7 +22,6 @@ class HenleggBehandlingController(
     private val tilgangService: TilgangService,
     private val henleggBehandlingService: HenleggBehandlingService,
     private val brevService: BrevService,
-    private val personopplysningerService: PersonopplysningerService,
 ) {
 
     @PostMapping("{behandlingId}/henlegg")
@@ -35,9 +29,7 @@ class HenleggBehandlingController(
         tilgangService.validerTilgangTilPersonMedRelasjonerForBehandling(behandlingId, AuditLoggerEvent.UPDATE)
         tilgangService.validerHarSaksbehandlerrolleTilStønadForBehandling(behandlingId)
         if (henlegg.skalSendeHenleggelsesbrev) {
-            validerIkkeSendBrevPåFeilType(henlegg)
-            validerIkkeSendBrevHvisVergemålEllerFullmakt(behandlingId)
-            brevService.opprettJournalførHenleggelsesbrevTask(behandlingId)
+            brevService.opprettJournalførHenleggelsesbrevTask(behandlingId, henlegg)
         }
         return Ressurs.success(henleggBehandlingService.henleggBehandling(behandlingId, henlegg))
     }
@@ -54,28 +46,4 @@ class HenleggBehandlingController(
     private fun henleggBrevRessurs(
         behandlingId: UUID,
     ) = Ressurs.success(brevService.genererHenleggelsesbrev(behandlingId))
-
-    private fun validerIkkeSendBrevPåFeilType(henlagt: HenlagtDto) {
-        feilHvis(henlagt.skalSendeHenleggelsesbrev && henlagt.årsak == HenlagtÅrsak.FEILREGISTRERT) { "Skal ikke sende brev hvis type er ulik trukket tilbake" }
-    }
-
-    private fun validerIkkeSendBrevHvisVergemålEllerFullmakt(
-        ident: UUID,
-    ) {
-        val personopplysninger = personopplysningerService.hentPersonopplysninger(ident)
-        val harVerge = personopplysninger.vergemål.isNotEmpty()
-        val harFullmakt: Boolean =
-            personopplysninger.fullmakt
-                .filter {
-                    it.gyldigTilOgMed == null ||
-                        (
-                            it.gyldigTilOgMed.isEqualOrAfter(
-                                LocalDate.now(),
-                            )
-                            )
-                }.isNotEmpty()
-        feilHvis(harVerge || harFullmakt) {
-            "Skal ikke sende brev hvis person er tilknyttet vergemål eller fullmakt"
-        }
-    }
 }
